+++ docs/release_notes.rst
@@ -12,7 +12,7 @@ introduce any backwards incompatible changes.
 django-wiki 0.2.2 (unreleased master branch)
 --------------------------------------------
 
- * Changes go here
+ * Remove ``wiki.decorators.json_view``, fixes server errors when resolving 404 links #604
 
 
 django-wiki 0.2.1